Added support for parsing teamcity build properties in templates.
When using custom parameters, you now have access to the TeamCity
build properties. 
This includes properties defined by the build configuration, as well
as env.*, teamcity.* and system.* properties.

Any property begining with the string "webhook." will automatically
be included in the POST payload. Others will not be included, but are
availble to the templating engine for use in custom messages.