Fix for GitHub issue #6 It appears that the implementation of the Branch interface available to an sBuild - and hence included in the webhook payload content - has references to other objects that end up being circular.
From limited testing, this appears to be only evident on non-default
branches of VCS that support multiple branches in TeamCity. For example,
feature branches on GIT or HG.

To avoid these references, now pass a simple Branch bean to the content
message constructed from the implementation passed to tcWebHooks by
TeamCity.