Process templates twice. The second time, use the output from the first parsing to build the template text.
This means one can now nest template variables.
eg,
<param name="body" value="${body.passed}" />

The first time, ${body.passed} will be populated and on the second
parsing, body will contain the text resolved in the first parsing.

The downside is that lots of warnings like the following:
java.lang.NoSuchMethodException thrown when trying to resolve value for
body.passed 
For that reason, these warnings have been downgraded to debug level in
log4j.