Added support for dates in templates. Use a variable that looks like ${now()} or ${now(yyyy-MM)} now() will use the default locale settings for displaying date now(yyyy-MM) will expect a SimpleDateFormat compatible text string representing the date format.
Also added a new getPreferredDateTimeFormat() method to the
WebHookTemplate interface. 
This puts a preferredDateFormat variable into the custom variables
list so that it is available to the template renderer.
Is that variable is present, the currentTime, buildStartTime and
buildFinishedTime values will be formatted using this string which is
passed into SimpleDateFormat.

This is currently utilised by the new ElasticSearchWebHookTemplate
because elastic search prefers dates in a specific format.

Also now() and now(simpleDateFormattedString) can be used in the URL.
This is also used in Elastic to that the webhook data can be POSTed to
an index with a date in it. For example:
http://myElasticServer:9200/teamcity-${now(yyyy-MM)}/buildevent