Create and register WebHookTemplateFileChangeHandler from Spring to fix #18
For some reason, creating the WebHookTemplateFileChangeHandler (with
new) from within the WebHookTemplateManager was getting another instance
of both classes at various times when loading the template list.

Getting spring to create the WebHookTemplateFileChangeHandler appears to
only create one instance of the class, and hence only one instance of
the template list.