Enable Preemptive Authentication by default, but allow disabling.
If a user has bothered to configured authentication credentials
then enabled sending those credentials straight away.

Don't wait for a 401 to be returned before sending the creds.

For those who are paranoid, it's possible to disable preemptive
authentication, and then the auth creds will not be sent if the REALM in
the 401 does not match the configured realm.