Isolate WebHookTemplate entity XML files from UI/REST API.
This is a large refactor to keep the WebHookTemplateEntity object from
leaking out of core.
The REST API and the WebUI should now be using WebHookTemplateConfig
items whenever they talk to core.

Also some optimisations when building the paylaod data.
Previously the code was finding the previous build repeatedly, which is
and expensive operation.