Add template content to TemplateItem requests
This is intended to be used by the Template Editor web page to allow
accessing all relevant information about a Build Event template with one
request.

Now one can access the API like so...

/app/rest/webhooks/templates/id:flowdock/templateItem/id:3?fields=content

...and the response back will include the webhook template content.

This is quite useful when asking for the following fields:

fields=useTemplateTextForBranch,href,content

which therefore includes all href's, content and the
useTemplateTextForBranch boolean flag.