Fix for failing webhooks with NullPointerException
There were a number of scenarios where a null would be returned if a
template or template content was not found. This was ok (not great) when
the templates were carefully edited, but now that one can edit/delete a
template from the API, it was possible to have a null returned in any of
the following scenarios:

- A webhook is configured, but the template has been removed.
- A webhook is configured, but the template no longer supports that
build event.
- A webhook is configured for a format type which is not valid (eg,
jsonblahblah rather than jsonTemplate).

Any of these would return a null, which was not handled correctly. This
would crash out of the loop and no more webhooks would be executed for
that buildState of that build instance.
Therefore, these errors were actually preventing valid webhooks from
getting the change execute.

Now these scenarios throw a WebHookContentResolutionException (or child
exception) and are handled in the loop.

They are written to a repository for display later.