Work around issue #112 by reinstating the missing TeamCity class.
The most recent version of the ExceptionMapperUtil file is extracted
from the SVN history, and copied into the project locally.

This is a temporary workaround to address
https://youtrack.jetbrains.com/issue/TW-56626 and
https://github.com/tcplugins/tcWebHooks/issues/112

The actual problem is caused by the removal of
src/jetbrains/buildServer/server/rest/jersey/ExceptionMapperUtil.java.

The replacement appears to be to extend ExceptionMapperBase instead,
however the ExceptionMapperBase class appears to force the response to
be text/plain, because it takes a ResponseData object and builds a
Jersey Response object itself. Previously, we were free to return our
own Response object, and could serialise the response as JSON or XML if
that was helpful.