Load stored webhook and then override with details from dialog.
Now when testing the webhook, if the webhook was previously saved, then
then saved webhook is loaded, and then the data from the webhook dialog
is overwritten.

This means that filters, parameters, etc are correctly loaded for
testing.

This commit also sends the auth data, so that authentication can be
tested even before the webhook is saved.