Set dialog height in CSS not JS. Fixes issue #123
Due to a change in Teamcity versions 2018.1.4 and 2018.2 RC, the
WebHookDialog, which extends BS.AbstractModalDialog, was no longer
setting the height correctly.

The JS code would run as part of ShowDialog, and once the dialog is
displayed, would read the height of the last tab and then set the height
of all other tabs to match. This resulted in the tabs all being uniform
height, so the dialog height would remain constant between tabs.

However, with the recent change in teamcity, the dialog was not actually
visible, so the measure phase returned 0. This set the height of the
tabs within the webhook dialog to 0px, and resulted in the text
overlfowing off the bottom of the dialog box.

The actual code to make the dialog visible must be being run after the
ShowDialog method completes now. It's being bound to the ShowDialog
finish event from what I can tell.

The workaround is to use CSS to set the min-height of the four tabs, and
remove the JS code.