Don't allow WebHook Preview and TestExecution to modify existing Configs
This addresses issue: #127

When a request is made to do a webhook preview or execution on an
existing webhook, the WebHookConfigFactoryImpl was returning the a
reference to the actual configuration from in memory. This object was
then modified and used for the preview or test webhook execution.

The changes in this object were not persisted (so the
plugin-settings.xml was not changed on disk), but subsequent page loads
would return the modified instance, rather then the persisted instance.

If subsequent edits were made to this instance, then the values from the
test execution would be loaded into the UI, and would effectively
over-write the persisted settings.

The solution is for the WebHookConfigFactoryImpl to return a copy of the
persisted configuration so that the WebHookUserRequestedExecutorImpl has
a temporary copy of the config to use for its executions.