Remove reference to payload format from plugin-settings.xml
Fixes issue #134

The reason for change this is because we now have a number template
formats, and templates should be able to change format without having to
update all the webhooks which reference that template.

Previously, the webhook configuration referred to a template and format
combination. However, templateId is unique, so referring to just
templateId allows a template to change format independently.

- Convert legacy formats to templates.
- Remove writing of payload format to config file.
- Create migration class to convert from format/template combo to
template.
- Remove references to templateFormatCombinationKey in the JS files.