Support for associating WebHook templates with Projects
This allows a WebHook Template to be created and associated with a
project rather then just _Root.
Now group admins can create templates for use with their own webhooks.
This addresses issue #131

Note: WebHook Template Ids must be unique across a teamcity instance,
even if the user creating the template is not aware and is not
permissioned to see an existing template with that ID.

- Uses templateId rather than templateName when referring to ID.
- Added projectId field and getProjectId method to templates.
- Added validators and permissions
- Fixed up permissions for template editing in REST API
- Added Templates to WebHook Project tab
- Added Templates list to Webhook Project Config tab.
- Added Project Name and link to Templates page.
- WebHooks now resolve project specific templates when building payload.