Add ability to limit or exclude vcs file list whilst building payload
This change includes the following new features:
1. The list of changed VCS files can be disabled (0).
2. The list of changed VCS files can be disabled if it is too large. The
default value is 100 files across all changes in a build.
3. The list can be always include (old behaviour) when the value is set
negative (-1).

NOTE: In the cases where the change list is disabled or too large (1 & 2
above), the payload will contain a `null` files list.
This is preferable to returning an empty list, because the empty list
implies no actual changed files were included in the change.
A null change list will typically be serialised to nothing in JSON or
XML, so the json array or xml element will be missing from the payload.
If your endpoint is expecting these, then it should fail in this
scenario or handle the missing field gracefully.

**Enables control over the maximum number of files changed in a build
before the changed file list is null.** This can be controlled in the
three ways below, in order of priority.

1. Add a 'param' named 'maxChangeFileListSize' to a webhook config by
editing plugin-settings.xml
2. Add a build parameter to a buildType or project called
'webhook.maxChangeFileListSize'
3. Define a property in teamcity's internal properties file called
'webhook.maxChangeFileListSize'

If the total number of files is greater than `maxChangeFileListSize`,
then the change list will be null. See note above.