Convert Java8 stream to For loop, since it's not compatible with TC9
TeamCity 9.x using Jersey 1.16, and asm 3.1

ASM 3.1, only supports Java7, so we can't use Java8 streams and such in
the REST API if we want to support TeamCity 9.

It appears that since TC10, Jetbrains have moved to using Jersey 1.19,
and a newer version of ASM which supports Java 8.

This is why this was never seen on TC10 and above. Fixes #156