Track secure value access and hide URL from UI
Keep a flag to determine when a secure: value is accessed. If so, hide
the full URL in the webUI so that secure values in the URL are not
leaked.

It's not possible to determine why the value was accessed. It could have
been to build the payload, or the URL. Therefore, if either of these
have resolved a secure value, we hide that URL.