Issue 53 hide password (#203)
* Update Velocity Templating Engine from 1.7 to 2.3
* Bump jetty-webapp in /tcwebhooks-core (#195)
* Format secure parameters correctly so that TeamCity obscures them

Previously, the value was prefixed with `secure:`. This was incorrect.
It is supposed to be a parameters name that is prefixed with `secure:`.

This fix changes that, so that TeamCity correctly swaps out the values
with tokens transparently.

* Track secure value access and hide URL from UI

Keep a flag to determine when a secure: value is accessed. If so, hide
the full URL in the webUI so that secure values in the URL are not
leaked.

It's not possible to determine why the value was accessed. It could have
been to build the payload, or the URL. Therefore, if either of these
have resolved a secure value, we hide that URL.

* Add support for hide-secure-values in the WebHookConfig

Adds ability to store/retreive the value of hide-secure-values.
This will be used as a flag to determine whether the UI and logs should
container secure values.

* Add support for hiding values from log and UI
* Show/Hide URL in UI when secure values checkbox is toggled.
* Populate external and internal project IDs when executing test
* Don't log Parameter values. They might contain a secure value.