Enable/disable stats and analytics. Refresh graph from UI
Includes updates for how the main-config.xml is written. It was missing
some settings.

Added buttons for enable/disable of statistics gathering (on by
default).
Added buttons for enable/disable of analytics reporting (off by
default).
Added a link that rebuilds the statistics xml files. This happens
hourly, but can now be requested from the UI.

All these actions reload that part of the page, and therefore redraw the
graph.