Remove build agent plugin
The build agent plugin is very trivial. It is really only there is
simplify sending a service message to trigger the service message event.

This can be achieved in other ways, and therefore removes the need for
Build Agents to download the plugin and restart themselves.

For large installations with many cloud agents, this
connect/download/restart/reconnect process happens for every agent and
some agents only do one build and then shutdown.