Remove references to method that was removed from TC REST API
Newer versions of the TC REST API now use a Java8 lambda supplier when
referencing a Value object.

Remove this pattern as it is no longer supported. We can't just change
to the supplier pattern because that might not be supported by older
versions of TeamCity.

Increment plugin version for release.