Add a deferred handler that rebuilds the webhook cache when setings load
We get an event whenever the plugin-settings.xml is reloaded on disk for
our element key <webhooks>.

It turns out that copying a project triggers this event on the new
project.

We can hook into this and queue a task to reload the cache for that
project. This is done with a deferred queue so that we give TC time to
reload the files from disk into the plugin settings cache.