Add support for BuildTypeDeleted
Because the buildType is deleted, we can't look it up to find the
internal Id when removing builds from a webhook config.

We have therefore added the baggage to the event, so that we get a copy
of the delete build config, and can use that to fetch the internal id.