Handle Project Persisted more rapidly. Map old BTs to new BTs
When a project is persisted, handle that event within 1 second.
This appears to be a better event to listen, as it looks like it fires
just after a project is written to disk.

What we're trying to do is handle the project copied pseudo event much
faster so that we can reload the webhook cache.

As part of this reload, also try to determine what the new bt ids might
be. When a webhook is persisted, it may container the buildTypeId of one
or more specific builds for which the webhook is configured.

When a teamcity project is copied, the webhooks are copied over too.
However, the buildTypeIds in the webhook will no longer reference valid
ids in the new project. We therefore try to map the old ids to the new
ones by doing a substring match on the buildType External Ids.