Handle case were WH is modified in a project we already have in cache
Previously, we ignored updates if we already knew about the project in
our map of project settings.

We previously expected that we'd be managing all updates. Of course this
doesn't work for VCS projects, so we need to react to the event when a
VCS change happens and we are already aware of the project, but probably
have a cache that is now out of sync.

Testing revealed that changes instigated from VCS can also leave ghost
webhooks around. This is because the cache was never cleaned for
webhooks that no longer exist.

Now the new list of webhooks in a project is compared against the list
of webhooks in the cache, and any in the cache that no-longer have a
matching webhook in the config are cleaned up.